Remove persistent storage
Fixes https://github.com/AndrewGable/ReactNativeChat/issues/15
